+++ test/dockerAssist.test.ts
@@ -529,10 +529,11 @@ function assertWORKDIR(item: CompletionItem, line: number, character: number, pr
 	assert.equal(item.textEdit.range.end.character, character + prefixLength);
 }
 
-function assertSourceImage(item: CompletionItem, sourceImage: string, documentation: string, startLine: number, startCharacter: number, endLine: number, endCharacter: number) {
+function assertSourceImage(item: CompletionItem, sourceImage: string, buildIndex: number, documentation: string, startLine: number, startCharacter: number, endLine: number, endCharacter: number) {
 	assert.equal(item.label, sourceImage);
 	assert.equal(item.kind, CompletionItemKind.Reference);
 	assert.equal(item.insertTextFormat, InsertTextFormat.PlainText);
+	assert.equal(item.sortText, buildIndex.toString());
 	assert.equal(item.documentation, documentation);
 	assert.equal(item.textEdit.newText, sourceImage);
 	assert.equal(item.textEdit.range.start.line, startLine);
@@ -1551,43 +1552,58 @@ describe('Docker Content Assist Tests', function() {
 				it("source image", function() {
 					var proposals = computePosition("FROM busybox AS source\n" + onbuild + "COPY --from=", 1, triggerOffset + 12);
 					assert.equal(proposals.length, 1);
-					assertSourceImage(proposals[0], "source", "busybox", 1, triggerOffset + 12, 1, triggerOffset + 12);
+					assertSourceImage(proposals[0], "source", 0, "busybox", 1, triggerOffset + 12, 1, triggerOffset + 12);
 				});
 	
 				it("source images alphabetical", function() {
 					var proposals = computePosition("FROM ubuntu:trusty AS setup\nFROM redhat:rawhide AS dev\n" + onbuild + "COPY --from=", 2, triggerOffset + 12);
 					assert.equal(proposals.length, 2);
-					assertSourceImage(proposals[0], "dev", "redhat:rawhide", 2, triggerOffset + 12, 2, triggerOffset + 12);
-					assertSourceImage(proposals[1], "setup", "ubuntu:trusty", 2, triggerOffset + 12, 2, triggerOffset + 12);
+					assertSourceImage(proposals[0], "setup", 0, "ubuntu:trusty", 2, triggerOffset + 12, 2, triggerOffset + 12);
+					assertSourceImage(proposals[1], "dev", 1, "redhat:rawhide", 2, triggerOffset + 12, 2, triggerOffset + 12);
 				});
 	
 				it("source image prefix", function() {
 					var proposals = computePosition("FROM busybox AS setup\nFROM busybox AS dev\n" + onbuild + "COPY --from=s", 2, triggerOffset + 13);
 					assert.equal(proposals.length, 1);
-					assertSourceImage(proposals[0], "setup", "busybox", 2, triggerOffset + 12, 2, triggerOffset + 13);
+					assertSourceImage(proposals[0], "setup", 0, "busybox", 2, triggerOffset + 12, 2, triggerOffset + 13);
 	
 					// casing should be ignored
 					proposals = computePosition("FROM busybox AS setup\nFROM busybox AS dev\n" + onbuild + "COPY --from=S", 2, triggerOffset + 13);
 					assert.equal(proposals.length, 1);
-					assertSourceImage(proposals[0], "setup", "busybox", 2, triggerOffset + 12, 2, triggerOffset + 13);
+					assertSourceImage(proposals[0], "setup", 0, "busybox", 2, triggerOffset + 12, 2, triggerOffset + 13);
 				});
 
 				it("no duplicate source images", function() {
 					let proposals = computePosition("FROM busybox AS source\nFROM busybox AS source\n" + onbuild + "COPY --from=", 2, triggerOffset + 12);
 					assert.equal(proposals.length, 1);
-					assertSourceImage(proposals[0], "source", "busybox", 2, triggerOffset + 12, 2, triggerOffset + 12);
+					assertSourceImage(proposals[0], "source", 0, "busybox", 2, triggerOffset + 12, 2, triggerOffset + 12);
 				});
 
 				it("no duplicate source images ignoring case", function() {
 					let proposals = computePosition("FROM busybox AS source\nFROM busybox AS soURCe\n" + onbuild + "COPY --from=", 2, triggerOffset + 12);
 					assert.equal(proposals.length, 1);
-					assertSourceImage(proposals[0], "source", "busybox", 2, triggerOffset + 12, 2, triggerOffset + 12);
+					assertSourceImage(proposals[0], "source", 0, "busybox", 2, triggerOffset + 12, 2, triggerOffset + 12);
 				});
 
 				it("only suggest previously declared source images", function() {
 					let proposals = computePosition("FROM node AS dev\n" + onbuild + "COPY --from=\nFROM node AS test", 1, triggerOffset + 12);
 					assert.equal(proposals.length, 1);
-					assertSourceImage(proposals[0], "dev", "node", 1, triggerOffset + 12, 1, triggerOffset + 12);
+					assertSourceImage(proposals[0], "dev", 0, "node", 1, triggerOffset + 12, 1, triggerOffset + 12);
+				});
+
+				it("source image indices", function() {
+					let proposals = computePosition("FROM alpine\nFROM node\n" + onbuild + "COPY --from=\nFROM busybox", 2, triggerOffset + 12);
+					assert.equal(proposals.length, 1);
+					assertSourceImage(proposals[0], "0", 0, "alpine", 2, triggerOffset + 12, 2, triggerOffset + 12);
+
+					proposals = computePosition("FROM alpine AS linux\nFROM node\nFROM busybox\n" + onbuild + "COPY --from=", 3, triggerOffset + 12);
+					assert.equal(proposals.length, 2);
+					assertSourceImage(proposals[0], "linux", 0, "alpine", 3, triggerOffset + 12, 3, triggerOffset + 12);
+					assertSourceImage(proposals[1], "1", 1, "node", 3, triggerOffset + 12, 3, triggerOffset + 12);
+
+					proposals = computePosition("FROM alpine AS 1\nFROM node\nFROM busybox\n" + onbuild + "COPY --from=", 3, triggerOffset + 12);
+					assert.equal(proposals.length, 1);
+					assertSourceImage(proposals[0], "1", 0, "alpine", 3, triggerOffset + 12, 3, triggerOffset + 12);
 				});
 			});